New file for app switch tests, add more tests
