+++ UnitTests/PayPalWebPaymentsTests/PayPalWebCheckoutClient_Tests.swift
@@ -4,7 +4,7 @@ import AuthenticationServices
 @testable import PayPalWebPayments
 @testable import TestShared
 
-// swiftlint: disable type_body_length file_length
+// swiftlint: disable type_body_length
 class PayPalClient_Tests: XCTestCase {
     
     var config: CoreConfig!
@@ -325,158 +325,6 @@ class PayPalClient_Tests: XCTestCase {
             checkoutURL,
             URL(string: "https://sandbox.paypal.com/checkoutnow?token=1234&redirect_uri=sdk.ios.paypal://x-callback-url/paypal-sdk/paypal-checkout&native_xo=1")
         )
-    }
-
-    // MARK: - handleReturnURL tests
-
-    func testHandleReturnURL_success_callsAppSwitchCompletionWithResult() {
-        var received: Result<PayPalWebCheckoutResult, CoreSDKError>?
-        payPalClient.appSwitchCompletion = { received = $0 }
-
-        let url = URL(string:
-            "https://appSwitchURL/success?token=ORDER123&PayerID=PAYER456&switch_initiated_time=1757431432185")!
-
-        payPalClient.handleReturnURL(url)
-
-        switch received {
-        case .success(let result)?:
-            XCTAssertEqual(result.orderID, "ORDER123")
-            XCTAssertEqual(result.payerID, "PAYER456")
-        default:
-            XCTFail("Expected success with PayPalWebCheckoutResult")
-        }
-
-        XCTAssertNil(payPalClient.appSwitchCompletion)
-    }
-
-    func testHandleReturnURL_cancel_mapsToCheckoutCanceledError() {
-        var received: Result<PayPalWebCheckoutResult, CoreSDKError>?
-        payPalClient.appSwitchCompletion = { received = $0 }
-
-        let url = URL(string:
-            "https://appSwitchURL/cancel?token=ORDER123&PayerID=PAYER456&switch_initiated_time=1757431432185"
-        )!
-
-        payPalClient.handleReturnURL(url)
-
-        if case .failure(let error)? = received {
-            XCTAssertTrue(PayPalError.isCheckoutCanceled(error))
-        } else {
-            XCTFail("Expected cancellation error")
-        }
-        XCTAssertNil(payPalClient.appSwitchCompletion)
-    }
-
-    func testHandleReturnURL_successPathMissingPayerID_isMalformedResultError() {
-        var received: Result<PayPalWebCheckoutResult, CoreSDKError>?
-        payPalClient.appSwitchCompletion = { received = $0 }
-
-        // Missing PayerID
-        let url = URL(string:
-            "https://appSwitchURL/success?token=ORDER123&switch_initiated_time=1757431432185"
-        )!
-
-        payPalClient.handleReturnURL(url)
-
-        if case .failure(let error)? = received {
-            XCTAssertEqual(error.code, PayPalError.malformedResultError.code)
-            XCTAssertEqual(error.domain, PayPalError.domain)
-        } else {
-            XCTFail("Expected malformedResultError")
-        }
-        XCTAssertNil(payPalClient.appSwitchCompletion)
-    }
-
-    func testHandleReturnURL_successPathIncorrectPayerIdFormat_isMalformedResultError() {
-        var received: Result<PayPalWebCheckoutResult, CoreSDKError>?
-        payPalClient.appSwitchCompletion = { received = $0 }
-
-        // Should be PayerID
-        let url = URL(string:
-            "https://appSwitchURL/success?token=ORDER123&PayerId=PAYER456&switch_initiated_time=1757431432185"
-        )!
-
-        payPalClient.handleReturnURL(url)
-
-        if case .failure(let error)? = received {
-            XCTAssertEqual(error.code, PayPalError.malformedResultError.code)
-            XCTAssertEqual(error.domain, PayPalError.domain)
-        } else {
-            XCTFail("Expected malformedResultError")
-        }
-        XCTAssertNil(payPalClient.appSwitchCompletion)
-    }
-
-    func testHandleReturnURL_successPathIncorrectPayeridFormat_isMalformedResultError() {
-        var received: Result<PayPalWebCheckoutResult, CoreSDKError>?
-        payPalClient.appSwitchCompletion = { received = $0 }
-
-        // Should be PayerID
-        let url = URL(string:
-            "https://appSwitchURL/success?token=ORDER123&Payer_id=PAYER456&switch_initiated_time=1757431432185"
-        )!
-
-        payPalClient.handleReturnURL(url)
-
-        if case .failure(let error)? = received {
-            XCTAssertEqual(error.code, PayPalError.malformedResultError.code)
-            XCTAssertEqual(error.domain, PayPalError.domain)
-        } else {
-            XCTFail("Expected malformedResultError")
-        }
-        XCTAssertNil(payPalClient.appSwitchCompletion)
-    }
-
-    func testHandleReturnURL__onlyCompletesOnce() {
-        var completionCount = 0
-        payPalClient.appSwitchCompletion = { _ in completionCount += 1 }
-
-        let url = URL(string:
-            "https://appSwitchURL/success?token=ORDER123&PayerID=PAYER456&switch_initiated_time=1757431432185"
-        )!
-
-        payPalClient.handleReturnURL(url)
-        // Second call should do nothing because appSwitchCompletion was cleared
-        payPalClient.handleReturnURL(url)
-
-        XCTAssertEqual(completionCount, 1, "Completion should be called exactly once")
-        XCTAssertNil(payPalClient.appSwitchCompletion)
-    }
-
-    func testHandleReturnURL_success_callsVaultAppSwitchCompletionWithResult() {
-        var received: Result<PayPalVaultResult, CoreSDKError>?
-        payPalClient.vaultAppSwitchCompletion = { received = $0 }
-
-        let url = URL(string:
-            "https://appSwitchURL/success?approval_token_id=345&approval_session_id=12345&PayerID=PAYER456&switch_initiated_time=1757431432185")!
-
-        payPalClient.handleReturnURL(url)
-
-        switch received {
-        case .success(let result)?:
-            XCTAssertEqual(result.approvalSessionID, "12345")
-            XCTAssertEqual(result.tokenID, "345")
-        default:
-            XCTFail("Expected success with PayPalVaultResult")
-        }
-
-        XCTAssertNil(payPalClient.vaultAppSwitchCompletion)
-    }
-
-    func testHandleReturnURL_onlyCompletesOnce_forVault() {
-        var completionCount = 0
-        payPalClient.vaultAppSwitchCompletion = { _ in completionCount += 1 }
-
-        let url = URL(string:
-            "https://appSwitchURL/success?approval_token_id=345&approval_session_id=12345&PayerID=PAYER456&switch_initiated_time=1757431432185"
-        )!
-
-        payPalClient.handleReturnURL(url)
-        // Second call should do nothing because vaultAppSwitchCompletion was cleared
-        payPalClient.handleReturnURL(url)
-
-        XCTAssertEqual(completionCount, 1, "Completion should be called exactly once")
-        XCTAssertNil(payPalClient.vaultAppSwitchCompletion)
     }
 }
 // swiftlint:enable type_body_length
